+++ src/commands/sk.rs
@@ -9,7 +9,7 @@ use std::{sync::Arc, thread};
 pub(crate) async fn invoke(
     query: &str,
     session_id: Option<String>,
-    session_start: Option<f32>,
+    session_start: Option<i64>,
 ) -> anyhow::Result<()> {
     let oshs = osh_files();
     // TODO filter here and in parallel?
